# specs/041-mt5-rest-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MT5 REST API
  description: |
    REST API for MetaTrader 5 data access. Provides read-only access to market data,
    account information, and trading history via HTTP endpoints.

    ## Features
    - **Multiple formats**: JSON and Apache Parquet responses
    - **Comprehensive data**: Symbols, ticks, rates, positions, orders, history
    - **Secure**: API key authentication
    - **Fast**: Async operations, optimized for high performance
    - **Documented**: Full OpenAPI 3.1 specification with examples

    ## Authentication
    All endpoints (except /health and /docs) require API key authentication via:
    - `X-API-Key` header

    ## Format Selection
    Choose response format via:
    - `Accept` header: `application/json` or `application/parquet`
    - `format` query parameter: `json` or `parquet`

    Default: JSON

  version: 1.0.0
  contact:
    name: pdmt5
    url: https://github.com/dceoy/pdmt5
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://localhost:8000
    description: Local development server
  - url: https://localhost:8443
    description: Local HTTPS server

security:
  - ApiKeyAuth: []

tags:
  - name: health
    description: Health check and system information
  - name: symbols
    description: Symbol information and market data
  - name: market
    description: Market data (rates, ticks)
  - name: account
    description: Account and terminal information
  - name: trading
    description: Positions and orders
  - name: history
    description: Historical orders and deals

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      description: Check API and MT5 terminal connection status
      operationId: getHealth
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /version:
    get:
      tags: [health]
      summary: Get MT5 version
      description: Get MetaTrader 5 terminal version information
      operationId: getVersion
      parameters:
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: MT5 version information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
            application/parquet:
              schema:
                type: string
                format: binary
        '503':
          $ref: '#/components/responses/MT5Disconnected'

  /symbols:
    get:
      tags: [symbols]
      summary: List symbols
      description: Get list of available trading symbols with optional filtering
      operationId: getSymbols
      parameters:
        - name: group
          in: query
          description: Symbol group filter (wildcards supported)
          schema:
            type: string
            example: "*USD*"
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
            application/parquet:
              schema:
                type: string
                format: binary
        '503':
          $ref: '#/components/responses/MT5Disconnected'

  /symbols/{symbol}:
    get:
      tags: [symbols]
      summary: Get symbol info
      description: Get detailed information for a specific symbol
      operationId: getSymbolInfo
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Symbol information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
        '404':
          $ref: '#/components/responses/SymbolNotFound'
        '503':
          $ref: '#/components/responses/MT5Disconnected'

  /symbols/{symbol}/tick:
    get:
      tags: [symbols]
      summary: Get latest tick
      description: Get last tick (price update) for a symbol
      operationId: getSymbolTick
      parameters:
        - $ref: '#/components/parameters/SymbolParam'
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Latest tick data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
        '404':
          $ref: '#/components/responses/SymbolNotFound'

  /rates/from:
    get:
      tags: [market]
      summary: Get rates from date
      description: Get historical OHLCV data starting from a specific date
      operationId: getRatesFrom
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            example: EURUSD
        - name: timeframe
          in: query
          required: true
          description: Timeframe in minutes
          schema:
            type: integer
            minimum: 1
            example: 60
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date-time
            example: "2024-01-01T00:00:00Z"
        - name: count
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            example: 100
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Historical rates data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
            application/parquet:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ValidationError'

  /rates/range:
    get:
      tags: [market]
      summary: Get rates for date range
      description: Get historical OHLCV data for a specific date range
      operationId: getRatesRange
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Historical rates data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /ticks/from:
    get:
      tags: [market]
      summary: Get ticks from date
      description: Get tick data starting from a specific date
      operationId: getTicksFrom
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: count
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100000
        - name: flags
          in: query
          description: "Tick flags: 2=INFO, 4=TRADE, 6=ALL"
          schema:
            type: integer
            default: 6
            enum: [2, 4, 6]
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Tick data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /account:
    get:
      tags: [account]
      summary: Get account info
      description: Get current account balance, equity, margin, and other account details
      operationId: getAccountInfo
      parameters:
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Account information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /terminal:
    get:
      tags: [account]
      summary: Get terminal info
      description: Get MetaTrader 5 terminal status and settings
      operationId: getTerminalInfo
      parameters:
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Terminal information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /positions:
    get:
      tags: [trading]
      summary: Get open positions
      description: Get currently open trading positions with optional filters
      operationId: getPositions
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: ticket
          in: query
          schema:
            type: integer
            minimum: 0
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Open positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /orders:
    get:
      tags: [trading]
      summary: Get pending orders
      description: Get currently pending orders with optional filters
      operationId: getOrders
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: ticket
          in: query
          schema:
            type: integer
            minimum: 0
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Pending orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /history/orders:
    get:
      tags: [history]
      summary: Get historical orders
      description: Get historical orders with date range or ticket/position filters
      operationId: getHistoryOrders
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
        - name: symbol
          in: query
          schema:
            type: string
        - name: ticket
          in: query
          schema:
            type: integer
        - name: position
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Historical orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

  /history/deals:
    get:
      tags: [history]
      summary: Get historical deals
      description: Get historical deals (executed trades) with filters
      operationId: getHistoryDeals
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
        - name: symbol
          in: query
          schema:
            type: string
        - name: ticket
          in: query
          schema:
            type: integer
        - name: position
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/FormatParam'
      responses:
        '200':
          description: Historical deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  parameters:
    SymbolParam:
      name: symbol
      in: path
      required: true
      description: Trading symbol name
      schema:
        type: string
        example: EURUSD

    FormatParam:
      name: format
      in: query
      description: Response format (json or parquet)
      schema:
        type: string
        enum: [json, parquet]
        default: json

  schemas:
    HealthResponse:
      type: object
      required: [status, mt5_connected, api_version]
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          example: healthy
        mt5_connected:
          type: boolean
          example: true
        mt5_version:
          type: string
          nullable: true
          example: "5.0.4321"
        api_version:
          type: string
          example: "1.0.0"

    DataResponse:
      type: object
      required: [data, count, format]
      properties:
        data:
          oneOf:
            - type: array
              items:
                type: object
              maxItems: 100000
            - type: object
          description: Response data (array for multiple records, object for single record)
        count:
          type: integer
          minimum: 0
          description: Number of records returned
          example: 100
        format:
          type: string
          enum: [json, parquet]
          example: json

    ErrorResponse:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          description: Error type URI
          example: "/errors/validation-error"
        title:
          type: string
          description: Short error summary
          example: "Request Validation Failed"
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Detailed error explanation
          example: "count must be positive (got: -10)"
        instance:
          type: string
          nullable: true
          description: Request URI that caused error
          example: "/rates/from"

  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "/errors/validation-error"
            title: "Request Validation Failed"
            status: 400
            detail: "count must be positive (got: -10)"
            instance: "/rates/from"

    SymbolNotFound:
      description: Symbol not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "/errors/invalid-symbol"
            title: "Symbol Not Found"
            status: 404
            detail: "Symbol 'INVALID' does not exist or is not available on this broker."
            instance: "/symbols/INVALID"

    MT5Disconnected:
      description: MT5 terminal not connected
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "/errors/mt5-disconnected"
            title: "MT5 Terminal Not Connected"
            status: 503
            detail: "MetaTrader5 terminal is not running or not logged in. Please start MT5 and login."
            instance: "/symbols"

    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "/errors/unauthorized"
            title: "Authentication Required"
            status: 401
            detail: "Missing or invalid API key. Provide X-API-Key header."
            instance: "/symbols"

    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: "/errors/rate-limit"
            title: "Rate Limit Exceeded"
            status: 429
            detail: "Too many requests. Limit: 100 requests per minute."
            instance: "/symbols"